Avoid dynamic linking against pyarrow
Dynamically linking against pyarrow doesn't work very well, as there are many scenarios where this doesn't work.
The straw that broke the camel's back was ray which includes a version of the pyarrow library itself (and doesn't coexist nicely with dlopening another version)
This change statically links some arrow C++ code so that we can generate in-memory buffers of the right format.  We then, in cython, take references to these buffers, and construct zero-copy PyArrow buffers that point to that memory.

I'm concerned there may be some memory issue, as there are several refcounting schemes involved in this dance (shared_ptr, cython refcount logic and our own PyObj refcounts), but the tests are consistently passing, and I'll keep an eye on it

+++ src/iters/get_index.hpp
@@ -76,7 +76,7 @@ namespace ss{ namespace iter{
         inline void read_sequence(const PyObj& obj) {
             Py_ssize_t  len = PySequence_Fast_GET_SIZE(obj.obj);
             for (auto &index : indexes) {
-                if (index.skip < len) { 
+                if (index.skip < (size_t)len) { 
                     (*index.destination) = PyObj(PySequence_Fast_GET_ITEM(obj.obj, index.skip));
                 } else {
                     (*index.destination) = UNDEFINED;
